# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /rbac/roles:
    post:
      operationId: RbacController_createRole
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleDto'
      responses:
        '201':
          description: 创建权限组(仅种子管理员)成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionGroup'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 创建权限组(仅种子管理员)
      tags:
        - rbac
      description: 创建权限组(仅种子管理员)。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
    get:
      operationId: RbacController_listRoles
      parameters: []
      responses:
        '200':
          description: 权限组列表(包含权限)成功。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PermissionGroup'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 权限组列表(包含权限)
      tags:
        - rbac
      description: 权限组列表(包含权限)。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
  /rbac/roles/{id}:
    patch:
      operationId: RbacController_updateRole
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleDto'
      responses:
        '200':
          description: 编辑权限组(仅种子管理员)成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionGroup'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 编辑权限组(仅种子管理员)
      tags:
        - rbac
      description: 编辑权限组(仅种子管理员)。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
    delete:
      operationId: RbacController_deleteRole
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: 删除权限组(仅种子管理员)成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletedResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 删除权限组(仅种子管理员)
      tags:
        - rbac
      description: 删除权限组(仅种子管理员)。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
  /rbac/permissions:
    post:
      operationId: RbacController_createPermission
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePermissionDto'
      responses:
        '201':
          description: 创建权限(仅种子管理员)成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Permission'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 创建权限(仅种子管理员)
      tags:
        - rbac
      description: 创建权限(仅种子管理员)。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
    get:
      operationId: RbacController_listPermissions
      parameters: []
      responses:
        '200':
          description: 列表:所有权限成功。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Permission'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 列表:所有权限
      tags:
        - rbac
      description: 列表:所有权限。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
  /rbac/permissions/{id}:
    delete:
      operationId: RbacController_deletePermission
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: 删除权限(仅种子管理员)成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletedResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 删除权限(仅种子管理员)
      tags:
        - rbac
      description: 删除权限(仅种子管理员)。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
  /rbac/roles/{roleId}/permissions:
    post:
      operationId: RbacController_attachPermissionFromBody
      parameters:
        - name: roleId
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachPermissionDto'
      responses:
        '201':
          description: 权限组配置权限(请求体形式,仅种子管理员)成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachedResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 权限组配置权限(请求体形式,仅种子管理员)
      tags:
        - rbac
      description: 权限组配置权限(请求体形式,仅种子管理员)。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
  /rbac/roles/{roleId}/permissions/{permissionId}:
    post:
      operationId: RbacController_attachPermission
      parameters:
        - name: roleId
          required: true
          in: path
          schema:
            type: number
        - name: permissionId
          required: true
          in: path
          schema:
            type: number
      responses:
        '201':
          description: 权限组配置权限(兼容 URL 形式,仅种子管理员)成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachedResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 权限组配置权限(兼容 URL 形式,仅种子管理员)
      tags:
        - rbac
      description: 权限组配置权限(兼容 URL 形式,仅种子管理员)。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
    delete:
      operationId: RbacController_detachPermission
      parameters:
        - name: roleId
          required: true
          in: path
          schema:
            type: number
        - name: permissionId
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: 权限组移除权限(仅种子管理员)成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetachedResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 权限组移除权限(仅种子管理员)
      tags:
        - rbac
      description: 权限组移除权限(仅种子管理员)。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
  /rbac/users/{userId}/roles:
    get:
      operationId: RbacController_listUserRoles
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: 用户已分配的权限组成功。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PermissionGroup'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 用户已分配的权限组
      tags:
        - rbac
      description: 用户已分配的权限组。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
    post:
      operationId: RbacController_assignRoleFromBody
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleDto'
      responses:
        '201':
          description: 用户分配权限组(请求体形式,仅种子管理员)成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignedResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 用户分配权限组(请求体形式,仅种子管理员)
      tags:
        - rbac
      description: 用户分配权限组(请求体形式,仅种子管理员)。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
  /rbac/users/{userId}/roles/{roleId}:
    post:
      operationId: RbacController_assignRole
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            type: number
        - name: roleId
          required: true
          in: path
          schema:
            type: number
      responses:
        '201':
          description: 用户分配权限组(兼容 URL 形式,仅种子管理员)成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignedResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 用户分配权限组(兼容 URL 形式,仅种子管理员)
      tags:
        - rbac
      description: 用户分配权限组(兼容 URL 形式,仅种子管理员)。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
    delete:
      operationId: RbacController_unassignRole
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            type: number
        - name: roleId
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: 用户移除权限组(仅种子管理员)成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnassignedResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - adminJwt: []
      summary: 用户移除权限组(仅种子管理员)
      tags:
        - rbac
      description: 用户移除权限组(仅种子管理员)。使用后台 JWT，并校验当前账号状态及对应 RBAC 权限。
...
info:
  title: R2RPC API
  description: R2RPC 管理控制面、调用方 RPC 与监控查询的 HTTP API。设备长连接协议使用 WebSocket，另见项目协议文档。
  version: 0.1.0
  contact:
    name: R2RPC Contributors
    url: https://github.com/rezoch340/R2RPC
    email: ''
  license:
    name: UNLICENSED
    url: https://github.com/rezoch340/R2RPC/blob/main/LICENSE
...
servers:
  - url: /
    description: 当前部署
components:
  securitySchemes:
    adminJwt:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: 后台登录接口签发的管理员 JWT。
    accessToken:
      scheme: bearer
      bearerFormat: R2RPC Access Token
      type: http
      description: 访问令牌管理页面生成的 rk_ 前缀 Access Token。
  schemas:
...
    CreateRoleDto:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        description:
          type: string
          maxLength: 255
      required:
        - name
    UpdateRoleDto:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        description:
          type: string
          maxLength: 255
    CreatePermissionDto:
      type: object
      properties:
        action:
          type: string
          minLength: 1
          maxLength: 64
        subject:
          type: string
          minLength: 1
          maxLength: 64
        description:
          type: string
          maxLength: 255
      required:
        - action
        - subject
    AttachPermissionDto:
      type: object
      properties:
        permissionId:
          type: number
          minimum: 1
      required:
        - permissionId
    AssignRoleDto:
      type: object
      properties:
        roleId:
          type: number
          minimum: 1
      required:
        - roleId
...
    ErrorResponse:
      type: object
      required:
        - statusCode
        - message
        - timestamp
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          oneOf:
            - type: string
              example: 请求参数非法
            - type: array
              items:
                type: string
              example:
                - name must be a string
            - type: object
              additionalProperties: true
              description: Nest HttpException 的结构化响应，例如 statusCode/message/error。
        timestamp:
          type: string
          format: date-time
    DeletedResult:
      type: object
      required:
        - deleted
      properties:
        deleted:
          type: boolean
          enum:
            - true
    AttachedResult:
      type: object
      required:
        - attached
      properties:
        attached:
          type: boolean
          enum:
            - true
    DetachedResult:
      type: object
      required:
        - detached
      properties:
        detached:
          type: boolean
          enum:
            - true
    AssignedResult:
      type: object
      required:
        - assigned
      properties:
        assigned:
          type: boolean
          enum:
            - true
    UnassignedResult:
      type: object
      required:
        - unassigned
      properties:
        unassigned:
          type: boolean
          enum:
            - true
    Permission:
      type: object
      required:
        - id
        - action
        - subject
      properties:
        id:
          type: integer
          example: 1
        action:
          type: string
          example: read
        subject:
          type: string
          example: monitor
        description:
          type: string
          nullable: true
        deletedAt:
          type: string
          format: date-time
          nullable: true
...
    PermissionGroup:
      type: object
      required:
        - id
        - name
        - createdAt
        - permissions
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: operator
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
...
  responses:
    BadRequest:
      description: 请求参数、路径参数或查询参数不符合契约。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 凭证缺失、无效、过期、撤销或对应账号已不存在。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 凭证有效，但账号、功能组或 RBAC 权限不允许本次操作。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 目标资源不存在或已被软删除。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 资源名称或关联关系与现有数据冲突。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: 服务端发生未预期错误。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'